+++ min_version_test/lib/main.dart
@@ -144,7 +144,7 @@ class _MyHomePageState extends State<MyHomePage> {
             Text(
               '$_counter',
               // ignore: deprecated_member_use
-              style: Theme.of(context).textTheme.headline4,
+              style: Theme.of(context).textTheme.headlineMedium,
             ),
           ],
         ),
